+++ packages/core/src/hooks/index.ts
@@ -26,3 +26,4 @@ export {SnapshotStateHook} from "./SnapshotStateHook";
 export {SupertestHook} from "./SupertestHook";
 export {ApplicationEventHook} from "./ApplicationEventHook";
 export {IocContainerHook} from "./IocContainerHook";
+export {ToxiproxyHook} from "./ToxiproxyHook";