+++ packages/node-test/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @nodeboot/node-test
 
+## 2.13.0
+
+### Minor Changes
+
+-   Add Toxiproxy test hook to provide network resilience test features
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/test@2.13.0
+
 ## 2.12.0
 
 ### Minor Changes